Moved expression rewrites to a module
