+++ rust/dyn-ddlog/src/ast/mod.rs
@@ -1,4 +1,5 @@
 mod debug_printing;
+pub mod rewrites;
 mod tree;
 
 #[cfg(feature = "debug-printing")]
